% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config}
\alias{config}
\title{Configure \code{odcr}}
\usage{
config(python = NA)
}
\arguments{
\item{python}{character, path to a python or conda executable that should be used by \code{odcr}. Note that the selected Python environment must have the \code{datacube} library installed (and its upstream Python and system dependencies must be available).}
}
\value{
None
}
\description{
This function configures \code{odcr} to connect with the correct Python environment and check the availability of the \code{datacube} Python library.
}
\examples{
\dontrun{
library(odcr)

config(python = "/env/bin/python3")
}
}
