% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xarray_methods.R
\name{dim}
\alias{dim}
\alias{dim.xarray.core.dataset.Dataset}
\alias{dim.xarray.core.dataarray.DataArray}
\title{Dimensions of \code{odcr} classes}
\usage{
\method{dim}{xarray.core.dataset.Dataset}(x)

\method{dim}{xarray.core.dataarray.DataArray}(x)
}
\arguments{
\item{x}{object of which dimensions should be retrieved.}
}
\value{
a (named) vector
}
\description{
\code{dim()} retrieves dimensions from an \code{xarray} object.
}
\examples{
\dontrun{
library(odcr)

# connect to a database, store the Daatcube connection internally (default and recommended)
database_connect(app = "Sentinel_2")

# build a query list
lat <- 22.821
lon <- 28.518
buffer <- 0.05

query <- list(
  'time' = c('2020-01', '2020-03'),
  'x' = c(lon - buffer, lon + buffer),
  'y' = c(lat + buffer, lat - buffer),
  'output_crs' = 'epsg:6933',
  'resolution' = c(-20,20)
)

# load data and return an xarray object for a query
ds <- dc_load(query = c(product = "s2_l2a", dask_chunks = dict(), query))

# return dimensions of ds
dim(ds)
}
}
