% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{xarray.core.dataarray.DataArray-class}
\alias{xarray.core.dataarray.DataArray-class}
\alias{xarray.core.dataarray.DataArray}
\title{"xarray.core.dataarray.DataArray" class}
\description{
"xarray.core.dataarray.DataArray" class
}
\concept{xarray.core.dataarray.DataArray}
