% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{xarray.core.dataset.Dataset-class}
\alias{xarray.core.dataset.Dataset-class}
\alias{xarray.core.dataset.Dataset}
\title{"xarray.core.dataset.Dataset" class}
\description{
"xarray.core.dataset.Dataset" class
}
\concept{xarray.core.dataset.Dataset}
