% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xarray_methods.R
\name{Arithmetic}
\alias{Arithmetic}
\alias{+.xarray.core.dataarray.DataArray}
\alias{-.xarray.core.dataarray.DataArray}
\alias{/.xarray.core.dataarray.DataArray}
\alias{*.xarray.core.dataarray.DataArray}
\title{Arithmetic operators for \code{odcr} classes}
\usage{
\method{+}{xarray.core.dataarray.DataArray}(xds, yds)

\method{-}{xarray.core.dataarray.DataArray}(xds, yds)

\method{/}{xarray.core.dataarray.DataArray}(xds, yds)

\method{*}{xarray.core.dataarray.DataArray}(xds, yds)
}
\arguments{
\item{xds}{\code{xarray} object}

\item{yds}{\code{xarray} object}
}
\value{
An \code{xarray} object of same class, shape and dimensions as the inputs to the operator, holding the result.
}
\description{
These arithmetic operators perform arithmetic on two objects of the same class, shape and dimensions.
}
