% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{as}
\alias{as}
\alias{coerce,xarray.core.dataset.Dataset,stars-method.}
\alias{coerce,xarray.core.dataarray.DataArray,stars-method.}
\alias{as.stars}
\alias{as.raster}
\title{Methods to coerce \code{odcr} classes to native spatial classes}
\usage{
as.stars(from)

as.raster(from)
}
\arguments{
\item{from}{object of class \code{xarray.core.dataset.Dataset}}
}
\value{
\code{\link[=as.stars]{as.stars()}} retruns an object of class \code{stars}

\code{\link[=as.raster]{as.raster()}} retruns an object of class \code{RasterLayer} or \code{RasterStack}
}
\description{
\code{\link[=as.stars]{as.stars()}} allows to convert \code{xarray.core.dataset.Dataset} to \code{stars}.

\code{\link[=as.raster]{as.raster()}} allows to convert \code{xarray.core.dataset.Dataset} to \code{RasterLayer} or \code{RasterStack}.
}
