% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xarray_methods.R
\name{dim}
\alias{dim}
\alias{dim.xarray.core.dataset.Dataset}
\alias{dim.xarray.core.dataarray.DataArray}
\title{Dimensions of \code{odcr} classes}
\usage{
\method{dim}{xarray.core.dataset.Dataset}(ds)

\method{dim}{xarray.core.dataarray.DataArray}(ds)
}
\arguments{
\item{ds}{dataset of which dimensions should be retrieved.}
}
\value{
a (named) vector
}
\description{
\code{dim()} retrieves dimensions from an \code{xarray} object.
}
