% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xarray_methods.R
\name{Extract}
\alias{Extract}
\alias{[.xarray.core.dataset.Dataset}
\alias{[.xarray.core.dataarray.DataArray}
\alias{[[.xarray.core.dataset.Dataset}
\alias{[[.xarray.core.dataarray.DataArray}
\alias{xar_sel_time}
\title{Methods to extract from \code{odcr} classes}
\usage{
\method{[}{xarray.core.dataset.Dataset}(ds, i = NULL, j = NULL, k = NULL, ...)

\method{[}{xarray.core.dataarray.DataArray}(ds, i = NULL, j = NULL, k = NULL, ...)

\method{[[}{xarray.core.dataset.Dataset}(ds, ...)

\method{[[}{xarray.core.dataarray.DataArray}(ds, ...)

xar_sel_time(ds, query, exact_match = T)
}
\arguments{
\item{ds}{\code{xarray} object, the dataset to be subsetted}

\item{i}{numeric, first dimension index}

\item{j}{numeric, second dimension index}

\item{k}{numeric, third dimension index}

\item{...}{numeric or character, index or indices by which to extract (additional) elements}

\item{query}{character vector, one or more time/date character vectors in the format of the time dimension of \code{ds} or an abbreviated form of it (e.g. only the date component)}

\item{exact_match}{logical, whether to return only exact matches (default) or to search for closest elements to each element in \code{query} using \verb{[difftime()]}}
}
\value{
Subsetted \code{xarray} object
}
\description{
\code{[} allows to subset an \code{xarray} object by its dimensions (see \code{\link[=dim]{dim()}})

\code{[[} allows to subset an \code{xarray} object by measurements/variables (e.g. spectral band), either named (character) or indexed (numeric).

\code{xar_sel_time()} allows to subset an \code{xarray} object by a character vector of dates/times.
}
