% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xarray_methods.R
\name{plot}
\alias{plot}
\alias{plot.xarray.core.dataset.Dataset}
\alias{plot.xarray.core.dataarray.DataArray}
\title{Plotting of \code{odcr} classes}
\usage{
\method{plot}{xarray.core.dataset.Dataset}(x, ...)

\method{plot}{xarray.core.dataarray.DataArray}(x, ...)
}
\arguments{
\item{x}{object which should be plotted}

\item{...}{additional arguments passed to \verb{[stars::plot()]}}
}
\value{
a starsplot
}
\description{
\code{plot()} plots an \code{xarray} object using \verb{[stars::plot()]}
}
