% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xarray_methods.R
\name{plot}
\alias{plot}
\alias{plot.xarray.core.dataset.Dataset}
\alias{plot.xarray.core.dataarray.DataArray}
\title{Plotting of \code{odcr} classes}
\usage{
\method{plot}{xarray.core.dataset.Dataset}(x, ...)

\method{plot}{xarray.core.dataarray.DataArray}(x, ...)
}
\arguments{
\item{x}{object which should be plotted}

\item{...}{additional arguments passed to \code{\link[stars:plot]{stars::plot()}}}
}
\value{
a \code{stars} plot
}
\description{
\code{plot()} plots an \code{xarray} object using \verb{[stars::plot()]}
}
\examples{
\dontrun{
library(odcr)

# connect to a database, store the Daatcube connection internally (default and recommended)
database_connect(app = "Sentinel_2")

# build a query list
lat <- 22.821
lon <- 28.518
buffer <- 0.05

query <- list(
  'time' = c('2020-01', '2020-03'),
  'x' = c(lon - buffer, lon + buffer),
  'y' = c(lat + buffer, lat - buffer),
  'output_crs' = 'epsg:6933',
  'resolution' = c(-20,20)
)

# load data and return an xarray object for a query
ds <- dc_load(query = c(product = "s2_l2a", dask_chunks = dict(), query))

# plot first timestamp, spatial subset and only first band/variable
plot(ds[1,101:300,101:300][[1]])
}
}
